refactor(bmn): DRY shared components untuk dokumen auction-candidates (#394)
Menghilangkan duplikasi besar antar-dokumen di modul BMN auction-candidates
dengan single source of truth, melanjutkan pola refactor kepegawaian (#392).

Surat Pernyataan (SPTJM, Nilai Limit, Tugas) — ~95% identik:
- _lib/print-pernyataan.ts: shared print CSS, buildPernyataanNomor, printPernyataan
- _components/PernyataanDocument.tsx: shared shell (KOP, judul, TTD) + PernyataanIdentity
- SptjmDocument 151->61, SptjLimitDocument 157->61, SpTugasDocument 133->47

SK documents (Penghentian, Panitia, Tim Penilai) — engine pagination ~230 baris
yang sebelumnya copy-paste 3x:
- _lib/sk-print.ts: runSkPagination(printWindow, config) terparameter prefix,
  section margin, label baris keputusan, dan kelas grup final
- SkPenghentian 1061->800, SkPanitia 893->638, SkTimPenilai 782->572

Perbaikan tampilan: isi baris "Menetapkan" pada MEMUTUSKAN kini bold di ketiga
SK (heading MEMUTUSKAN tetap bold, label Menetapkan tetap normal).

Zero behavior change pada output cetak/PDF. CSS statis tiap dokumen tetap inline.
Validasi: ESLint --max-warnings=0, tsc --noEmit, next build 59/59 bersih.
